requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1.2
idna==2.10
mysqlclient==2.0.1
Pillow==8.0.0
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10
